requirements.txt
// Copyright 2017 TsumiNa. All rights reserved.
// Use of this source code is governed by a BSD-style
// license that can be found in the LICENSE file.

numpy
scipy
pandas
ruamel.yaml
tqdm
jupyter
jupyterlab
matplotlib